Changed so that a new interval is created everytime one presses a button. Also no direction change on the same axis.
Tried the interval-thing to see if it would change direction more smoothly, but it's still not there yet.
Same-axis-direction: when the snake moves up, it should not do anything when pressing up or down. Same goes for left/right, in this case it should only change direction when pressing up or down. Only change if it is on the other axis.