Added a pseudocode-file with a plan for how to implement a gameBoard array
The gameBoard array will separate the game logic from the actual graphics. The array is basically a x/y-coordinate system. This will introduce
a separate function for updating the GUI which always reads and redraws a board every interval. The gameBoard array consists of tile-objects
which have x/y-coordinates and a flag which could be set to "tile", "snake", "apple" or anything else.