Fixed offset bug where the snake is drawn below the gameboard
Placing a snake body item in y-coordinate 0 would draw the snake one game pixel below the board instead of on the first row. This was due
to rectangles in canvas being drawn downwards, while I have chosen to use a traditional y/x-axis grid which goes up and to the right
(instead of down and to the right). Seems more logical this way. Fixed it by setting the y-offset to minus one game pixel when drawing the snake.